requirements.txt
numpy>=1.24.0
opencv-python>=4.7.0
Pillow>=10.0.0
pygame>=2.3.0
